Build
  - linux_gnu: changed default optimization to -O3 -ffast-math -funroll-loops
  - linux_gnu: added -fPIC (required for Ubuntu 18.04(?))
  - linux_gnu: changed papi home from /usr/local to /usr

Mesh

  - removed misleading Index::is_zero()
  - removed Block::rank() in favor of cello::rank()

PPML

  - Updated loop bounds in PPML potok[x-z] to be 3,n[xyz]-3 not 1,n[xyz]-1
    on recommendation from Alexei Kritusk

Enzo

  - added enzo::block_array()
  - changed

Parameters

  - removed solver_[min|max]_level_coarse; added solver_coarse_level

Cello

  - use cello::rank() globally to access rank

Solvers

  - added debug code to solvers
  - bug fixes to BiCgStab, Cg, and Mg0
  - consolidated EnzoSolverMg0::[min|max]_level_coarse_ into coarse_level_
  - restructured EnzoSolverMg0 to somewhat clarify control flow

Tools

  - updated perf-summary.awk to include zone compute rates

+++ src/Cello/problem_Problem.cpp
@@ -795,7 +795,7 @@ Physics * Problem::create_physics_
 
 Physics * Problem::physics (std::string type) const throw()
 {
-  for (int i=0; i<physics_list_.size(); i++) {
+  for (size_t i=0; i<physics_list_.size(); i++) {
     if (physics_list_[i]->type() == type) return physics_list_[i];
   }
   return NULL;
